requirements.txt
absl-py==2.1.0
agate==1.9.1
aiodebug==2.3.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.6
aiosignal==1.3.1
alembic==1.13.3
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.0
apache-airflow==2.9.3+composer
apache-airflow-providers-amazon==8.29.0
apache-airflow-providers-apache-beam==5.8.0
apache-airflow-providers-celery==3.8.2
apache-airflow-providers-cncf-kubernetes==8.4.2
apache-airflow-providers-common-compat==1.2.0
apache-airflow-providers-common-io==1.4.1
apache-airflow-providers-common-sql==1.17.0
apache-airflow-providers-dbt-cloud==3.10.1
apache-airflow-providers-fab==1.2.2
apache-airflow-providers-ftp==3.11.1
apache-airflow-providers-google==10.23.0
apache-airflow-providers-hashicorp==3.8.0
apache-airflow-providers-http==4.13.1
apache-airflow-providers-imap==3.7.0
apache-airflow-providers-mysql==5.7.1
apache-airflow-providers-postgres==5.13.0
apache-airflow-providers-sendgrid==3.6.0
apache-airflow-providers-smtp==1.8.0
apache-airflow-providers-sqlite==3.9.0
apache-airflow-providers-ssh==3.13.1
apache-beam==2.59.0
apispec==6.6.1
appdirs==1.4.4
argcomplete==3.5.0
asgiref==3.8.1
asn1crypto==1.5.1
astunparse==1.6.3
attrs==24.2.0
babel==2.16.0
backoff==2.2.1
backports.tarfile==1.2.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
billiard==4.2.1
blinker==1.8.2
boto3==1.35.36
botocore==1.35.36
CacheControl==0.14.0
cachelib==0.9.0
cachetools==5.5.0
cattrs==24.1.2
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
clickclick==20.10.2
cloudpickle==2.2.1
colorama==0.4.6
colorlog==4.8.0
ConfigUpdater==3.2
connexion==2.14.2
crcmod==1.7
cron-descriptor==1.4.5
croniter==3.0.3
cryptography==43.0.1
daff==1.3.46
db-dtypes==1.3.0
dbt-adapters==1.7.0
dbt-bigquery==1.8.2
dbt-common==1.10.0
dbt-core==1.8.7
dbt-extractor==0.5.1
dbt-semantic-interfaces==0.5.1
decorator==5.1.1
deepdiff==7.0.1
Deprecated==1.2.14
diff_cover==9.2.0
dill==0.3.1.1
distlib==0.3.8
dnspython==2.6.1
docopt==0.6.2
docstring_parser==0.16
docutils==0.21.2
email_validator==2.2.0
fastavro==1.9.7
fasteners==0.19
filelock==3.13.3
firebase-admin==6.5.0
Flask==2.2.5
Flask-AppBuilder==4.5.0
Flask-Babel==2.0.0
Flask-Bcrypt==1.0.1
Flask-Caching==2.3.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.8.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
flatbuffers==24.3.25
flower==2.0.1
frozenlist==1.4.1
fsspec==2024.9.0
gast==0.6.0
gcloud-aio-auth==5.3.2
gcloud-aio-bigquery==7.1.0
gcloud-aio-storage==9.3.0
gcsfs==2024.9.0.post1
google-ads==25.0.0
google-analytics-admin==0.23.0
google-api-core==2.20.0
google-api-python-client==2.147.0
google-apitools==0.5.32
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-access-context-manager==0.2.1
google-cloud-aiplatform==1.68.0
google-cloud-appengine-logging==1.4.5
google-cloud-asset==3.26.4
google-cloud-audit-log==0.3.0
google-cloud-automl==2.13.5
google-cloud-batch==0.17.28
google-cloud-bigquery==3.25.0
google-cloud-bigquery-datatransfer==3.15.7
google-cloud-bigtable==2.26.0
google-cloud-build==3.25.0
google-cloud-common==1.3.5
google-cloud-compute==1.19.2
google-cloud-container==2.51.0
google-cloud-core==2.4.1
google-cloud-datacatalog==3.20.1
google-cloud-datacatalog-lineage==0.3.1
google-cloud-datacatalog-lineage-producer-client==0.1.0
google-cloud-dataflow-client==0.8.12
google-cloud-dataform==0.5.11
google-cloud-dataplex==2.2.2
google-cloud-dataproc==5.12.0
google-cloud-dataproc-metastore==1.15.5
google-cloud-datastore==2.20.1
google-cloud-dlp==3.23.0
google-cloud-documentai==2.32.0
google-cloud-filestore==1.9.5
google-cloud-firestore==2.19.0
google-cloud-kms==3.0.0
google-cloud-language==2.14.0
google-cloud-logging==3.11.2
google-cloud-memcache==1.9.5
google-cloud-monitoring==2.22.2
google-cloud-orchestration-airflow==1.14.0
google-cloud-org-policy==1.11.1
google-cloud-os-config==1.17.5
google-cloud-os-login==2.14.6
google-cloud-pubsub==2.23.1
google-cloud-pubsublite==0.6.1
google-cloud-redis==2.15.5
google-cloud-resource-manager==1.12.5
google-cloud-run==0.10.8
google-cloud-secret-manager==2.20.2
google-cloud-spanner==3.49.1
google-cloud-speech==2.27.0
google-cloud-storage==2.18.2
google-cloud-storage-transfer==1.12.0
google-cloud-tasks==2.16.5
google-cloud-texttospeech==2.17.2
google-cloud-translate==3.16.0
google-cloud-videointelligence==2.13.5
google-cloud-vision==3.7.4
google-cloud-workflows==1.14.5
google-crc32c==1.6.0
google-pasta==0.2.0
google-re2==1.1.20240702
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpc-google-iam-v1==0.13.1
grpc-interceptor==0.15.4
grpcio==1.66.1
grpcio-gcp==0.2.6
grpcio-health-checking==1.66.1
grpcio-status==1.66.1
gspread==6.0.0
gunicorn==20.1.1
h11==0.14.0
h5py==3.10.0
hdfs==2.8.3
